custom permission added in expense api

+++ accounts/views.py
@@ -12,41 +12,49 @@ from django.db import IntegrityError
 class ExpenseCreateAPIView(APIView):
     permission_classes = [IsAuthenticated]
     def post(self, request, *args, **kwargs):
-        data = request.data
+        user_groups = request.user.groups.values_list('name', flat=True)  # Get the names of all user groups
+        has_perm = False
+        for i in user_groups:
+            if i == 'CFO':
+                has_perm = True
 
-        try:
-            subject = data['subject']
-            amount = data['amount']
-            month = data['month']
-            currency = data['currency']
+        if has_perm:
+            data = request.data
 
-            # Check if an expense with the same details already exists
-            existing_expense = Expense.objects.filter(
-                subject=subject,
-                amount=amount,
-                user_id=request.user.id,
-                month=month,
-                currency=currency
-            ).first()
+            try:
+                subject = data['subject']
+                amount = data['amount']
+                month = data['month']
+                currency = data['currency']
 
-            if existing_expense:
-                return Response({'error': 'Expense with the same details already exists'}, status=status.HTTP_400_BAD_REQUEST)
+                # Check if an expense with the same details already exists
+                existing_expense = Expense.objects.filter(
+                    subject=subject,
+                    amount=amount,
+                    user_id=request.user.id,
+                    month=month,
+                    currency=currency
+                ).first()
 
-            # Create Expense object
-            expense = Expense.objects.create(
-                subject=subject,
-                amount=amount,
-                user_id=request.user.id,
-                month=month,
-                currency=currency
-            )
+                if existing_expense:
+                    return Response({'error': 'Expense with the same details already exists'}, status=status.HTTP_400_BAD_REQUEST)
 
-            return Response({'id': expense.id,'subject':expense.subject,'amount':expense.amount,'month':expense.month,'currency':expense.currency,'user':expense.user.email,'created_at':expense.created_at}, status=status.HTTP_201_CREATED)
+                # Create Expense object
+                expense = Expense.objects.create(
+                    subject=subject,
+                    amount=amount,
+                    user_id=request.user.id,
+                    month=month,
+                    currency=currency
+                )
 
-        except KeyError as e:
-            return Response({'error': f'Missing required field: {e}'}, status=status.HTTP_400_BAD_REQUEST)
+                return Response({'id': expense.id,'subject':expense.subject,'amount':expense.amount,'month':expense.month,'currency':expense.currency,'user':expense.user.email,'created_at':expense.created_at}, status=status.HTTP_201_CREATED)
+
+            except KeyError as e:
+                return Response({'error': f'Missing required field: {e}'}, status=status.HTTP_400_BAD_REQUEST)
     
-
+        else:
+            return Response({'error': 'You don\'t have permission to enter expense.'}, status=status.HTTP_403_FORBIDDEN)
 
     def get(self, request):
         # Get parameters from the request
@@ -116,44 +124,62 @@ class ExpenseUpdateAPIView(APIView):
 
     
     def put(self, request, expense_id):
-        # Retrieve the expense instance
-        try:
-            expense = Expense.objects.get(id=expense_id)
-            data = request.data
+        user_groups = request.user.groups.values_list('name', flat=True)  # Get the names of all user groups
 
-            attributes_to_update = [
-                'subject', 'amount','currency', 'created_at', 'month'
-            ]
+        has_perm = False
+        for i in user_groups:
+            if i == 'CFO':
+                has_perm = True
 
-            for attribute in attributes_to_update:
-                if attribute in data:
-                    setattr(expense, attribute, data[attribute])
+        if has_perm:
+            try:
+                expense = Expense.objects.get(id=expense_id)
+                data = request.data
 
-            expense.save()
+                attributes_to_update = [
+                    'subject', 'amount','currency', 'created_at', 'month'
+                ]
 
-            expense_data = {
-                    'id': expense.id,
-                    'subject': expense.subject,
-                    'amount': expense.amount if expense.amount else None,
-                    'currency': expense.currency if expense.currency else None,
-                    'created_at': expense.created_at if expense.created_at else None,
-                    'month': expense.month if expense.month else None,
-                    'user': expense.user.email
+                for attribute in attributes_to_update:
+                    if attribute in data:
+                        setattr(expense, attribute, data[attribute])
 
-                }
+                expense.save()
 
-            return Response(expense_data, status=status.HTTP_200_OK)
-        except Exception as e:
-            return Response({'error': f'{e}'}, status=status.HTTP_400_BAD_REQUEST)
-    
+                expense_data = {
+                        'id': expense.id,
+                        'subject': expense.subject,
+                        'amount': expense.amount if expense.amount else None,
+                        'currency': expense.currency if expense.currency else None,
+                        'created_at': expense.created_at if expense.created_at else None,
+                        'month': expense.month if expense.month else None,
+                        'user': expense.user.email
+
+                    }
+
+                return Response(expense_data, status=status.HTTP_200_OK)
+            except Exception as e:
+                return Response({'error': f'{e}'}, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({'error': 'You don\'t have permission to update expense.'}, status=status.HTTP_403_FORBIDDEN)
 
     def delete(self, request, expense_id):
-        try:
-            expense = Expense.objects.get(pk=expense_id)
-            expense.delete()
-            return JsonResponse({'message': 'Expense deleted successfully'}, status=200)
-        except Expense.DoesNotExist:
-            return JsonResponse({'error': 'Expense not found'}, status=404)
-        except IntegrityError:
-            return JsonResponse({'error': 'IntegrityError occurred while deleting expense'}, status=500)
+        user_groups = request.user.groups.values_list('name', flat=True)  # Get the names of all user groups
+
+        has_perm = False
+        for i in user_groups:
+            if i == 'CFO':
+                has_perm = True
+
+        if has_perm:
+            try:
+                expense = Expense.objects.get(pk=expense_id)
+                expense.delete()
+                return JsonResponse({'message': 'Expense deleted successfully'}, status=200)
+            except Expense.DoesNotExist:
+                return JsonResponse({'error': 'Expense not found'}, status=404)
+            except IntegrityError:
+                return JsonResponse({'error': 'IntegrityError occurred while deleting expense'}, status=500)
+        else:
+            return Response({'error': 'You don\'t have permission to delete expense.'}, status=status.HTTP_403_FORBIDDEN)
